+++ ExoGraphGUI/internal/functions/Connect-ExoGraphGuiService.ps1
@@ -56,7 +56,7 @@
     Process {
         # Connect to Graph if there is no current context
         $conn = Get-MgContext
-        $requiredScopes = "Mail.ReadWrite", "Mail.Send", "MailboxSettings.Read"
+        $requiredScopes = "Mail.ReadWrite", "Mail.Send", "MailboxSettings.Read", "User.ReadWrite.All"
         if ( $conn ) {
             $compare = Compare-Object -ReferenceObject $conn.scopes -DifferenceObject $requiredScopes -IncludeEqual
         }